feat: Windows builds
